Add CNAME file on deploy and handle gh-pagees branch existing locally
